Move aws test setup to conftest
To make it resusable
