+++ tests/conftest.py
@@ -5,6 +5,7 @@ import uuid
 from contextlib import contextmanager
 from typing import Optional
 
+import boto3
 import pytest
 from cpr_sdk.embedding import Embedder
 from cpr_sdk.search_adaptors import Vespa, VespaSearchAdapter
@@ -24,6 +25,49 @@ from app.core import security
 from app.core.aws import S3Client, get_s3_client
 from app.db.session import get_db
 from app.main import app
+
+
+@pytest.fixture(scope="function")
+def mock_aws_creds():
+    """Mocked AWS Credentials for moto."""
+    os.environ["AWS_ACCESS_KEY_ID"] = "test"
+    os.environ["AWS_SECRET_ACCESS_KEY"] = "test"
+    os.environ["AWS_SECURITY_TOKEN"] = "test"
+    os.environ["AWS_SESSION_TOKEN"] = "test"
+
+
+@pytest.fixture(scope="function")
+def mock_pipeline_bucket(mock_aws_creds):
+    pipeline_bucket_name = "test_pipeline_bucket"
+    ingest_trigger_root = "input"
+    os.environ["PIPELINE_BUCKET"] = pipeline_bucket_name
+    os.environ["INGEST_TRIGGER_ROOT"] = ingest_trigger_root
+
+    with mock_s3():
+        mock_s3_client = boto3.client("s3", region_name=os.getenv("AWS_REGION"))
+        mock_s3_client.create_bucket(
+            Bucket=pipeline_bucket_name,
+            CreateBucketConfiguration={
+                "LocationConstraint": os.getenv("AWS_REGION"),
+            },
+        )
+
+        test_prefixes = [
+            "2024-15-22T21.53.26.945831",
+            "2024-01-02T18.10.56.827645",
+            "2023-12-10T23.11.27.584565",
+            "2023-07-15T14.33.31.783564",
+            "2022-11-06T14.57.17.873576",
+            "2022-05-03T15.38.21.245423",
+        ]
+        for prefix in test_prefixes:
+
+            mock_s3_client.put_object(
+                Bucket=pipeline_bucket_name,
+                Key=f"{ingest_trigger_root}/{prefix}/test_file.txt",
+                Body="data".encode(),
+            )
+        yield pipeline_bucket_name
 
 
 @pytest.fixture
